+++ src/Phan/Plugin/Internal/UseReturnValuePlugin.php
@@ -54,6 +54,7 @@ class UseReturnValuePlugin extends PluginV3 implements PostAnalyzeNodeCapability
     public const UseReturnValueInternal = 'PhanPluginUseReturnValueInternal';
     public const UseReturnValueInternalKnown = 'PhanPluginUseReturnValueInternalKnown';
     public const UseReturnValueNoopVoid = 'PhanPluginUseReturnValueNoopVoid';
+    public const UseReturnValueGenerator = 'PhanPluginUseReturnValueGenerator';
     // phpcs:enable Generic.NamingConventions.UpperCaseConstantName.ClassConstantNotUpperCase
 
     public const DEFAULT_THRESHOLD_PERCENTAGE = 98;