+++ NEWS.md
@@ -21,6 +21,9 @@ New features (Analysis):
   New issue type: `PhanCommentOverrideOnNonOverrideProperty`.
   (Phan already warns for constants and methods)
 
+Plugins:
++ Emit `PhanPluginUseReturnValueGenerator` for calling a function returning a generator without using the returned Generator. (#4013)
+
 Bug fixes:
 + Properly analyze the right hand side for `$cond || throw ...;` (e.g. emit `PhanCompatibleThrowException`) (#4199)
 + Don't infer implications of `left || right` on the right hand expression when the right hand side has no side effects. (#4199)